First portion of reading obj wavefront file
Store material of obj file and print the material name.
The diffuse texture is read from the mtl.
Note that the jpg diffuse texture must be in the same directory as the model for it to find it. Didn't create a way to handle a texture that is missing and will most likely panic the program.