Ran model loading test
Ran some test by dragging different models in the application some passed and some failed.
created notes on what cause the application to fail and panic.
Note that the goal for the application is to handle all the panics and not force the application to terminate.
there are some small difference between the original model and the model loaded in  by the application (Joker) got to figure what caused this abnormality.