change the math crate to glam. and refactor and optimized camera.
change math crate from cgmath to glam since it was faster in a math bench test and utilizes a more stable SIMD library.

Thing I would like to improve on the camera is to switch the rotation from using euler as rotation to quaternions and also compact variable such as position, rotation, scale to a homogenous transformation matrix. These are improvement that I have to learn in order to implement them.