Created a blueprint  for the lighting system.
Currently the code for the application is quite messy and need to be refactored and redesigned (material system, lighting system, model system, and the graphic system), but I'm currently experimenting with the lighting system for my application.

The system only support one light source and only direction light. There a a lot of limitation and messy or incorrect calculation which will be refined and polish as I understand lighting more.

Also the material system now read specular color, diffuse color, ambient color, shininess, dissolve, and optical density from the mtl and apply their data to the wgsl shader. Except optical density