replaced rotation from Euler angle to quaternion. Small cleaning on the application which is still in process
Converted camera rotation from using Euler angle to populate the rotation matrix to Quaternion which add the current quaternion to the offset quaternion to get the desire quaternion and populated the rotation matrix with the rotation matrix that has been created by the desired quaternion.
Also did some cleaning on the application. The cleaning is still not complete and the quaternion is still quite not complete as well.