completed texture module and clean dependency
Completed the necessity for the texture module. and clean up all the scripts that I've been working on in the rendering module.  Clean up dependency for rendering and made it more easy to reason about.

there are still issues within texture module that I will re-look at such as compression and transcoding basis file which currently causes a heap corruption.