improving performance for icosphere before continuing on the implementation
Change the Hasher for the hashmap since it was slowing down the function for generating the ico sphere. Switch to a very fast hashing algorithm that isn't Dos safe nor cryptographic safe. multiplied radius and normalized the starting vertices from the ico sphere rather then iterating over it and manipulating the data in chucks . I will optimize middlePoint function and then probably unfold it to remove the function all together. Then i will continue with retrieving the rest of the data for the ico-sphere such as the normal, uv, tangent, bi-tangent.