Finished Icosphere and other updates.
Finished the Icosphere primitive. Currently there is an issue with the uv zigzag artifact when the uv is mapping from 0.9 to 0.1 which squishes the 0.9 vertices to 0.1 vertices giving a zig zag artifact. Going to comeback to find a solution to this artifact.
 Here the visual issue
https://i.stack.imgur.com/gcAz8.png

Note removed all the calculation for tangent and bi-tangent, since some developer will not need it. I still kept both tangent and bitangent [f32; 4] in Vertex, but developer can calculate Recalulate Tangent using the function in mesh_util (calculate tangents) I will also create a function to calculate bitangent (binormal). This will make the creation of primitive faster since calculating tangent and bitangent in primitive are removed completely and easy function call is exposed to developer that require the tangent or bitangent.

Removed unsafe function clamp_ss, min_ss, max_ss