Creating mesh utility script.
Creating mesh utility script that will generate the vertices normal, tangent or bi-tangent. There will be different way on generating the normal for the mesh depending on the user. Eg. for generating the normal the user can chose to make the normal flat or to smooth the normals. There will be customization ability for tangent or bi-tangent if possible. All features to help ease user creating or calculating mesh data will be in the mesh_util.rs file