Added more test and working on the the material
Added more test to the overall project and continuing the development of the material module of the game engine.
The material module should work with the Shader module where it reflect the property of the shader and from there the material module should deduce the variable type and set a default value to the variable in a mtrl file.