Relooking and redoing material_parser in shader module
Relooking at material_parser in shader module. Completely breaking it and going to re-build it from the ground up with the material system in place.
The material parser should be able to read the supported shader and retrieve all variable with binding group and store it in a mtrl file. The mtrl file should be customizable like the unity material file and should have value for each binding variable. It will then be fed back the the render system and the render system will fed the value to the correct binding group variable, thus updating the variable in the shader.