Flattened shader struct member data
Flattened shader struct member data into a single collection with a key to hold retrieve to member data from the single collection and a identifier to let the data within the single collection know what struct it belongs.
The identifier is the struct binding and group in the shader which is required to pass the data to the shader.