Create a two method for material_parser.
Create two method that where we can retrieve all the
(Set (GLSL), group (WGSL)) variable with extra metadata that matches the index number passed into the function. We can also retrieve all the binding variable with extra metadata that matches the index number passed into the other function.

Going to continue design the architecture for material module which is tied to the shader module, so there will be alot of refactoring and improving in the shader module and rethinking on the design choice.