Removed Vec chunk struct in favour for ByteArray in core.
Working on design and implementation on ByteArray in core to work with data that are Pod (plain old data struct) and zeroable, but not in struct format to produce the same bytes array as data within a struct. Experimenting with the architecture design, so think will definitely change and be improved.