Continuing on the material module.
Just about finished the material module. Going to relook at the whole module once more before continuing to the next module which would be either lighting or camera.
One thing I think can be improved would be the material_target.rs script. The big MaterialTarget enum which contains Wrapper of the supported primitives' type in the shader. This will get extremely big as i add more primitive type that are supported by this game engine and it will also get bigger in size. Currently it is 80 bytes which is below my expectation of less then or equal to 64 bytes.