Removed shader reflection
Removed shader reflection script. It looked more like a feature that isn't needed currently. I moved the code source to a local folder. If there is any need for shader reflection I will add it to this engine code source until then it will be staying in the "Pit" of unused code features.