Writing idea comment of the layout and implementation
Writing idea comment in the camera module that will help me flesh out the code source for the camera module. The camera will not have one mega component like some game engine (Viewport, Camera matrix, Projection, etc..). The engine is built on ecs so it will have smaller component that are added to an entity that will camera because of the specific component attached to the entity.