Working progress on the light conversion
Working on the light conversion. Support EV to luminance conversion. There will be more conversion support to help  end user (people working on the engine) have tool to do physically base lighting and for the engine to easily convert measurement to calculate the display the result to the front user (using the engine).