tested raw_clip and raw_ambisonic_clip script
tested  raw_clip and raw_ambisonic_clip  script and re-looked at the script. It looks simple and good to continue to the next script. I have added test to show the design of the script and make sure everything is running correctly currently and in the future when making changes.