audio recording very early preview
currently implementing audio recording. It only save to wav format currently and will eventually support other format. The audio sound quite high so I got to look at it and it need to be refactored and re-design.

Currently it is a proof of working stage and will have drastic changes.