Major refactor and optimizing.
Major refactor and optimizing for the progress so far on the audio module.  Improved readability and performance. Just have to look over the life time later, since everything is static currently. This will change in the future.  The Audio module is just about complete there are just some core features that are missing.

Going defer Midi, DSP, and HRTF for the future.
Got to really think about if I should implement HRTF, since it quite computationally expensive to compute and is a "nice to have feature".