Added function for quaternion construction & removed redundant component
Added function for quaternion construction & removed redundant component.

Rotation matrix to quaternion,
Euler to quaternion,
angle axis to quaternion,
angle axis magnitude to quaternion,

I have removed space, orientation and other scripts / components. The next step is to remove WorldToLocal and build the engine without it and stub it in if needed. Going to only keep LocalToWorld and better naming for variables and re-organize the module.

I have also improved some calculation the the transform module.

Going to make it a strict discipline to make the variable name descriptive and add test before implementation.