cleaning up previous implementation of source code.
cleaning up source code and refactoring to map it to the ECS vision I got for the game engine.
Currently working progress on Fabled_render module.
also re-naming variables so it is expression on what it is used for and what it stores. Following the rule where you re-write your code three time, but in my case a little more then three times.