Working progress on color module
Applying what I am learning.
Currently worked on color transform for sRgb and hdr (RGBE)
going to added retrieving HSV  from a specified color.
Note that the value will be clamp from zero to a value that can be greater then one to support hsv to hdr.
Thinks will change a lot in the color folder