implementing lua binding for audio crate
I am implementing a designing the way the binding will works as well as the api for lua early on, so I can spot architectural and design flaws for both rust api  and lua api and the binding that binds them.

I am working on the audio module binding.

Rollback to shipyard 0.5.0 since the newer version is not yet stable and I don't really have a solution of how to implement the trait required by shipyard to be used for the system. I don't want to import shipyard and it's dependency for each module.