calculate perspective with fov h and fov y
calculate camera perspective takes into account if the field of view is horizontal or if it is vertical. Create perspective with focal length where it is in mm. Not yet been tested. Use the fov mapping function to convert focal to fov with explicit parameters for compute_perspective_matrix function.