optimized camera.rs script
happy with camera.rs script the only thing that needs to be changed is compute_look_at_matrix fn the +Z is back I want it to be forward. That the standard in the game engine +X is Right, +Y is Up and +Z is forward. After this, I will work on the aspect ratio.