working on fov algorithm
Working progress on the field of view algorithm.

Removed both vertical minus and stretched.

vertical minus reduces fov when aspect ratio increases which will be problematic for dual monitors.
As for stretched if the fov and aspect ratio are not calibrated then there will be significant distortion. Horizontal plus is standardized in-game while Anamorphic is for fine control.