generalizing function and fleshing out grading
chromatic adaption now take a tristimulus value and a adaption parameter which will allow the function to take in different color space and continuing to flesh out color grading. Almost done, note that this will not be advance. If you need some advance grading use davinci resolve or other similar application. Don't want to re-create the wheel and it will bloat and complicate the source code.  After that I will re-work on the color blending script. Then create the necessary component for this color module.