blend function done.
Finished blend function. Going to add the required component in the color folder then optimize then that should be a wrap. Added Idea for when I work on the lighting. IES parser and possible Eulumdat parser. Added on to bool vector math container. Clean math a little to remove warnings.