Done with spot light shadow  shader for now. Going for point light next
Point light will be not be stored as a cube map nor Parabloid or Dual-Parabloid since the trade off doesn't align with the vision I have to this game engine. We will store the 6 textures in a single texture strip that will be the texture width, but will be the texture height * 6.0 for the six sides. We can then add a offset to sample the next texture which will be on the sample 2d texture map. So point light shader will have a texture_depth_2d_array. Where the array is for supporting multiply point light and a single texture_depth_2d will contain all six sides of th texture in that one texture.